+++ app/infrastructure/services/__init__.py
@@ -10,7 +10,6 @@ from infrastructure.services.dependencies import (
     MaxMindClientDep,
     EventDispatcherDep,
     TranslationServiceDep,
-    IdempotencyServiceDep,
 )
 from infrastructure.services.providers import (
     get_app_settings,
@@ -18,7 +17,6 @@ from infrastructure.services.providers import (
     get_maxmind_client,
     get_event_dispatcher,
     get_translation_service,
-    get_idempotency_service,
     t,
 )
 
@@ -29,14 +27,12 @@ __all__ = [
     "MaxMindClientDep",
     "EventDispatcherDep",
     "TranslationServiceDep",
-    "IdempotencyServiceDep",
     # Core providers
     "get_app_settings",
     "get_settings",
     "get_maxmind_client",
     "get_event_dispatcher",
     "get_translation_service",
-    "get_idempotency_service",
     # Translation helper
     "t",
 ]
